injest hooks class
